Add points to the Hilbert curve

Add points to the Hilbert curve

# S4 method for HilbertCurve
hc_segmented_points(object, ir = NULL, x1 = NULL, x2 = NULL, gp = gpar(),
    np = max(c(2, 10 - hc_level(object))),
    mean_mode = c("w0", "absolute", "weighted", "max_freq"),
    shape = "circle")

Arguments

object

A HilbertCurve-class object.

ir

an IRanges object which specifies the input intervals.

x1

if start positions are not integers, they can be set by x1.

x2

if end positions are not integers, they can be set by x2.

np

number of points (a circle or a square, ...) that are put in a segment.

gp

graphic parameters for points. It should be specified by gpar. The size of the points can be set here because the size of points are determined by np argument.

mean_mode

when a segment in the curve overlaps with intervals in ir, how to calculate the mean values for this segment. See explanation in hc_points.

shape

shape of points. Possible values are "circle", "square", "triangle", "hexagon", "star".

Details

Every segment on the curve is split by np points.

This function is used internally, please use hc_points,HilbertCurve-method directly.

Value

A data frame which contains coordinates (in the 2D space) of points.

Author

Zuguang Gu <z.gu@dkfz.de>

Examples

# see documentation of hc_points
NULL
#> NULL